move mesh module out of the data_models submodule
