+++ src/dfastbe/bank_erosion/data_models/inputs.py
@@ -7,7 +7,7 @@ from typing import ClassVar, Dict, List, Tuple
 import numpy as np
 from dfastio.xyc.models import XYCModel
 from shapely.geometry import LineString
-from dfastbe.bank_erosion.mesh import MeshData
+from dfastbe.bank_erosion.mesh.mesh_data import MeshData
 from dfastbe.bank_erosion.data_models.calculation import (
     SingleBank,
     SingleLevelParameters,
